moved cict app perf common fixtures to common utilities
